+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of cookbook-openstack-obj
 * Upgrading berkshelf from 2.0.18 to 3.1.5
 * Bump Chef gem to 11.16
 * Add keystone registration support
+* Add swift.conf template
 
 ## 9.0.3
 * Bugfix run_command exitstatus